Version 0.5.2
Bug fix in recordProductCounts route where the employee id was not being read in during the access control verification process.
Now sends employee permission levels in the login response.
Now shows the product name in the getAllLatestProductCounts response.
Removes the ability to overwrite the employee id from the original product count record when updating the record.