Version 0.6.2
Fixes a bug in the CheckEmplyoeePin route which fails to check if the employee ID entered exists in the database.

Time routes now return a Boolean value to represent whether or not an employee already took/started a break to make it easier for the frontend to render the appropriate options to users.

Now, the clock out and break start/end routes clear the access token cookie to make sure that employees are forced to sign in after every action on the time system. Note that this occurs even in error cases.