Version 0.6.9
This is the initial test of BMS-Suite on Docker.

Environment variable files and their locations remain unchanged, but values may need to be updated. If testing locally, the DB_HOST values should be changed from 'localhost' to 'host.docker.internal' if applicable.

Note: This version's backend server can now only be run in Docker since the static frontend file locations are different in the container. If the server needs to be run outside Docker, you must manually copy the contents of /frontend/build to /backend/src/static

TODO: Update README.md and test production stability